refactor a componentes y hooks para mejor entendimiento
