+++ components/map/truck/icon.tsx
