+++ components/map/warehouse/warehouse-component.tsx
@@ -1,10 +1,10 @@
-import { Circle, Text, Image, Rect, Group, Line, Path, Star } from "react-konva";
-import useCalRoute from "@/hooks/useCalRoute";
-import { useMapContext } from "@/contexts/MapContext";
+import { TooltipType } from "@/components/map/tooltip/tooltip-component";
+import { useMapContext } from "@/contexts/ContextMap";
+import { useTrucksCount } from "@/hooks/use-trucks-count";
+import useCalRoute from "@/hooks/use-callroute";
 import { KonvaEventObject } from "konva/lib/Node";
 import { Dispatch, SetStateAction } from "react";
-import { TooltipType } from "@/components/map/tooltip/MapTooltip";
-import { useTrucksCount } from "@/hooks/use-trucks-count";
+import { Circle, Group, Path, Rect, Star, Text } from "react-konva";
 
 interface StoreHouseProps {
   posX: number;
