+++ components/planner/results.tsx
