+++ components/simulations/results.tsx
