+++ components/simulations/setup.tsx
@@ -40,9 +40,9 @@ import { format } from "date-fns";
 import { Label } from "../ui/label";
 import { es } from "date-fns/locale";
 import { Badge } from "@/components/ui/badge";
-import BlockService from "@/services/block.service";
-import PedidosService from "@/services/pedidos.service";
-import { useSimulationContext } from "@/contexts/SimulationContext";
+import BlockService from "@/services/blockages.service";
+import PedidosService from "@/services/orders.service";
+import { useSimulationContext } from "@/contexts/ContextSimulation";
 import { Calendar as CalendarComponent } from "@/components/ui/calendar";
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
 import { SimulationInterface, SimulationType } from "@/interfaces/simulation.interface";
