+++ components/simulations/status.tsx
