+++ components/trucks/details.tsx
